artifacts/imports/ia_clinique_starter/ia_clinique_starter/requirements.txt

pyyaml
jsonschema
markdown-it-py
requests
pytest
pre-commit
